chat: route unknown agent models to zai like agents-page test; model_server removal, sandbox and codex auth updates

+++ crates/agent_3th_cli/ai_interface_layer/src/request.rs
@@ -3,13 +3,13 @@ use crate::message::Message;
 /// Hard cap on messages sent in one request (prompt size guard).
 pub const MAX_MESSAGES: usize = 64;
 /// Neutral sampling temperature used when a request doesn't specify one.
-pub const DEFAULT_TEMPERATURE: f32 = 0.7;
+pub const DEFAULT_TEMPERATURE: f64 = 0.7;
 
 #[derive(Debug, Clone)]
 pub struct ChatRequest {
     pub model: String,
     pub messages: Vec<Message>,
-    pub temperature: f32,
+    pub temperature: f64,
 }
 
 impl ChatRequest {
@@ -26,7 +26,7 @@ impl ChatRequest {
         Self::new(model, vec![Message::user(prompt)])
     }
 
-    pub fn with_temperature(mut self, temperature: f32) -> Self {
+    pub fn with_temperature(mut self, temperature: f64) -> Self {
         self.temperature = temperature;
         self
     }
